Fix bug for paused timers
The buggy logic for pausing and playing was that
whenever an active timer was rendered, the timer
would be played in the setup, and the play function
controlled the intervals for updating. This caused any
paused timers to be played when switching pages.
Now, play is not called in the setup and the intervals
are moved to the add function when timers are created.